requirements.txt
anyio==4.9.0
asttokens==3.0.0
attrs==24.3.0
Automat==24.8.1
beautifulsoup4==4.13.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipython==8.18.1
itemadapter==0.10.0
itemloaders==1.3.2
jedi==0.19.2
jmespath==1.0.1
lxml==4.9.4
matplotlib-inline==0.1.7
numpy==2.0.2
packaging==24.2
pandas==2.2.3
parsel==1.9.1
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
Protego==0.3.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-telegram-bot==22.0
pytz==2025.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
requests-futures==1.0.2
Scrapy==2.12.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
tldextract==5.1.3
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
w3lib==2.2.1
wcwidth==0.2.13
yahooquery==2.3.7
zope.interface==7.2
